layout/style/test/test_bug771043.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=771043
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 771043</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 771043 **/
  var expectedValue;
  var callCount = 0;
  var storedHeight;
  function callback(arg) {
    ++callCount;
    is(arg.matches, expectedValue,
       "Should have the right value on call #" + callCount + " to the callback");
  }

  function flushLayout() {
    storedHeight = document.querySelector("iframe").offsetHeight;
  }

  function setHeight(height, reason) {
    var oldCount = callCount;
    var ifr = document.querySelector("iframe");
    ifr.style.height = height + "px";
    flushLayout();
    is(callCount, oldCount+1, "Should have called our callback when " + reason);
  }

  SimpleTest.waitForExplicitFinish();
  addLoadEvent(function() {
    var mql = frames[0].matchMedia("(orientation: landscape)");
    mql.addListener(callback);
    expectedValue = true;
    setHeight(50, "going landscape");
    expectedValue = false;
    setHeight(200, "going portrait");

    var ifr = document.querySelector("iframe");
    ifr.style.display = "none";
    flushLayout();
    ifr.style.display = "";
    expectedValue = true;
    setHeight(50, "going landscape from display:none");

    expectedValue = false;
    setHeight(200, "going portrait after display:none");
    SimpleTest.finish();
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=771043">Mozilla Bug 771043</a>
<!-- Important: the iframe needs to be displayed -->
<p id="display"><iframe style="width: 100px; height: 200px"</p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
